% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCoordinates.R
\name{addCoordinates}
\alias{addCoordinates}
\title{Adding the longitude and latitude coordinate to the input data set when the (Belgian) postal code is known}
\usage{
addCoordinates(inputDT, dirShape = NULL)
}
\arguments{
\item{inputDT}{data.table object containing the data of interest. This is an obligatory argument, without default value.}

\item{dirShape}{Character vector of length 1 indicating directory where all the shape files are stored. This is an obligatory argument, with NULL as default value.}
}
\value{
The longitude and latitude coordinates are added as 2 separate columns to the 'inputDT' argument. Note that no output object is thrown by the 'addCoordinates' function, the 2 columns are directly added to the object that was passed to the 'inputDT' argument.
}
\description{
Adding the longitude and latitude coordinate to the input data set when the (Belgian) postal code is known
}
\examples{
nSamples <- 50000
observedValues <- rpois(nSamples, lambda = 1)
library(data.table)
inputDT <- as.data.table(data.frame(exposure = round(runif(nSamples), digits = 2),
observed = observedValues,
postalCode = sample(c('1000', '3000', '5000'), nSamples, replace = TRUE)))
addCoordinates(inputDT)
#addCoordinates(inputDT, '/home/robin/bla/')
}
