% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binningUnivSplineClassInterval.R
\name{binningUnivSplineClassInterval}
\alias{binningUnivSplineClassInterval}
\title{Get the spline splits based on the specific style of splitting based on the [classInt] package.}
\usage{
binningUnivSplineClassInterval(
  splineEst,
  nGroups = NULL,
  sampleSizeBin = 50000,
  nSamples = 10,
  verbose = TRUE,
  style = "fisher"
)
}
\arguments{
\item{splineEst}{number with the spline estimates of rht}

\item{nGroups}{positive integer, 
The number of splits in the y-direction. 
Note that this can result in more groups if you want to split in the x-direction}

\item{sampleSizeBin}{positive integer, 
the number of sample to be taken for the splits}

\item{nSamples}{positive integer, 
the number of random samples that needs to be taken for the splits}

\item{verbose}{`logical(1)` if the status should be shown or not}

\item{style}{style for the function [classIntervals()]}
}
\value{
splits in the y-direction
}
\description{
Get the spline splits based on the specific style of splitting based on the [classInt] package.
}
