% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binningUnivSplineTree.R
\name{binningUnivSplineTree}
\alias{binningUnivSplineTree}
\title{Split the GAM samples to the best splits based on a tree-model of the `rpart` package. 
Bootstrap will be used for the sampling of the trees. 
For the final breaks are the mediods of all samples, using the function [cluster::pam()]}
\usage{
binningUnivSplineTree(
  inputDT,
  nGroups,
  sampleSizeBin = min(nrow(inputDT), 50000),
  nSamples = 10,
  maxIter = 500
)
}
\arguments{
\item{inputDT}{data.table object with columns y and x for the tree-model. 
y is the target variable and x the variable which needs to be splitted}

\item{nGroups}{positive integer, number of groups that needs to be created}

\item{sampleSizeBin}{positive integer, datasize used for the tree-model}

\item{nSamples}{positive integer, number of bootstrapped samples that needs to be created}

\item{maxIter}{positive integer, maximum number of iterations for the tree model. 
The complexity parameter `cp` from the `rpart` package will be used to have the proper number of groups}
}
\value{
list of the final break points
}
\description{
Split the GAM samples to the best splits based on a tree-model of the `rpart` package. 
Bootstrap will be used for the sampling of the trees. 
For the final breaks are the mediods of all samples, using the function [cluster::pam()]
}
