% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateExplainedVariance.R
\name{calculateExplainedVariance}
\alias{calculateExplainedVariance}
\title{Calculate the Explained Variance}
\usage{
calculateExplainedVariance(
  inputDT,
  splits,
  splineEst,
  varName,
  groupedVarName = sprintf("\%sGrouped", varName)
)
}
\arguments{
\item{inputDT}{data.table with the input data}

\item{splits}{Named list of numeric vectors
The names of this object need to correspond to column names of the 'inputDT' argument. 
This is an obligatory argument, without default value.}

\item{splineEst}{named list of the spline estimates of the GAM model}

\item{varName}{column name of `splits` to get the estimates from}

\item{groupedVarName}{column based on `varName` which is already grouped}
}
\value{
single numeric value, which is the total explained variance
}
\description{
Calculate the Explained Variance
}
