% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Check-Functions.R
\name{checkWholeNumb}
\alias{checkWholeNumb}
\title{Checking if all elements of a list are all numeric or integer vectors with just whole numbers}
\usage{
checkWholeNumb(listNum, namesListElements = NULL)
}
\arguments{
\item{listNum}{A list of the vectors of which one wishes to check if their data type is numeric}

\item{namesListElements}{Character vector containing the names of the variables of which the data type is checked. Optional parameter, with as default value NULL. 
This argument should be used when the variable of which the data type is checked is not an object that was provided as an argument to the function, 
or when the list elements of the first argument do not have a name attached to it.}
}
\value{
No value is returned if all vectors have the numeric or integer data type, containing just whole numbers. 
If not, an error message is thrown for each element of the list that does not pertain to the numeric data type.
}
\description{
Checking if all elements of a list are all numeric or integer vectors with just whole numbers
}
\examples{
arg1 <- 2
checkWholeNumb(list(arg1))
checkWholeNumb(list(2))

arg2 <- 0.8
\donttest{checkWholeNumb(list(arg2))}
\donttest{checkWholeNumb(list(1, arg2))}
}
