% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractSplineEstimate.R
\name{extractSplineEstimate}
\alias{extractSplineEstimate}
\title{Extract the Spline estimates of the GAM model}
\usage{
extractSplineEstimate(gamFit, keyWordColSelection = NULL)
}
\arguments{
\item{gamFit}{object gam, with the already runned GAM model of [mgcv::gam()]}

\item{keyWordColSelection}{list with the smooth terms to be extracted from the GAM. 
If NULL all the smooth terms will be used}
}
\value{
list of length of the `keyWordColSelection` containing a list with all the estimates in the GAM. 
The number of lines will then be the number of rows of the given dataset.
}
\description{
Extract the Spline estimates of the GAM model
}
