% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBestnGroups.R
\name{getBestnGroups}
\alias{getBestnGroups}
\title{Select the best splits based on the explained variance}
\usage{
getBestnGroups(
  explVar,
  nGroups,
  splittingPoints,
  minIncr = 0.05,
  minIncrRel = 1.05
)
}
\arguments{
\item{explVar}{list with all the explained variances}

\item{nGroups}{list with all the explained nGroups}

\item{splittingPoints}{list with all the splitting points.}

\item{minIncr}{single numeric value. 
Indicating the minimum increase of the explained Variance}

\item{minIncrRel}{single numeric value.
Indicating the minimum relative increase of the explained variance}
}
\value{
list with the best solutions of length 3 with names: 
`ExplainedVariance`: best explained variance
`nGroups`: Number of groups with had the best explained variance
`SplittingPoints`: poisitions of the best splits
}
\description{
Select the best splits based on the explained variance
}
