% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optNumbGroups.R
\name{optNumbGroups}
\alias{optNumbGroups}
\title{Find the optimal number of splits using a GAM model as a base}
\usage{
optNumbGroups(
  inputDT,
  gamFit,
  splineEst,
  splineVarNames,
  nGroups = seq(4, 10, 2),
  typeSplit = "tree",
  sampleSizeBin = min(50000, nrow(inputDT)),
  minIncr = 0.05,
  minIncrRel = 1.05,
  minExplVar = 0.6,
  verbose = TRUE,
  nSamples = 10,
  maxIter = 500,
  invLink = exp,
  style = "fisher"
)
}
\arguments{
\item{inputDT}{data.table used for the GAM model to fit}

\item{gamFit}{fitted GAM model}

\item{splineEst}{list of the estimated spline models}

\item{splineVarNames}{list of (continuous) variable names which should be binned. 
all the names should be found in the `names(splineEst)` and in  `inputDT`}

\item{nGroups}{list of integers, 
the different sizes to look for the binning.}

\item{typeSplit}{"tree" or "classInt"
"tree" being the split based on the trees
"classInt" using the [classIntervals()] function to split. 
a vector can also be given. 
The length should be the same asl the length of `splineVarNames`.}

\item{sampleSizeBin}{integer, 
sample size which should be used for the binning strategy of a tree}

\item{minIncr}{single numeric value. 
Indicating the minimum increase of the explained Variance}

\item{minIncrRel}{single numeric value.
Indicating the minimum relative increase of the explained variance}

\item{minExplVar}{single numeric value, 
what the minimum explained variance should be. 
If the minimum explained variance is not met a warning will be given}

\item{verbose}{logical, 
if current status of the modeling should be shown}

\item{nSamples}{integer, 
number of different trees that needs to be constructed to find the best splits}

\item{maxIter}{maximum number of iterations for the tree construction}

\item{invLink}{function, the inverse of the link function. 
Usually this is the function [exp()]}

\item{style}{see `style` in function [classIntervals()].
Default being `fisher`}
}
\value{
list with the explained variance and the best splits for each variable given.
}
\description{
Find the optimal number of splits using a GAM model as a base
}
