% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingSpatialEffect.R
\name{plottingSpatialEffect}
\alias{plottingSpatialEffect}
\title{Constructing a ggplot-styled map of the Belgium, splitted up by INS region, to represent a zone by means of a continuous variable or simply the exposure}
\usage{
plottingSpatialEffect(
  inputDT,
  predictions = NULL,
  plotTitle = "",
  dirShape = NULL,
  plotDir = NULL,
  plotName = NULL,
  pdf2 = TRUE,
  lowColor = NULL,
  highColor = NULL,
  predAsFact = FALSE,
  predFactColors = NULL,
  randomPick = FALSE
)
}
\arguments{
\item{inputDT}{data.table object containing the data of interest. Note that the 'inputDT' argument needs to contain an 'exposure ' and 'postalCode' column in order to the function to work. This is an obligatory argument, without default value.}

\item{predictions}{Numeric vector of length corresponding to the number of rows of the 'inputDT' argument. This argument correspond to the predictions (mostly the spline covariate effect estimates) made for every observation of the 'inputDT' argument. In case that the NULL value is supplied to this argument, the exposure is plotted. This is an optional argument, without default value.}

\item{plotTitle}{Character vector of length 1 consisting of the plot title of the plot. This argument needs to be supplied in conjunction with the 'plotDir' argument. This is an optional argument, with '' as default value.}

\item{dirShape}{Character vector of length 1 consisting of the directory where one can find the shape files. This is an optional argument, with NULL as default value.}

\item{plotDir}{Character vector of length 1 consisting of the directory where one wants to save the ggplot object to. This argument needs to be supplied in conjunction with the 'plotName' argument. This is an optional argument, with NULL as default value.}

\item{plotName}{Character vector of length 1 consisting of the file name of the saved ggplot object to. This argument needs to be supplied in conjunction with the 'plotDir' argument. This is an optional argument, with NULL as default value.}

\item{pdf2}{Logical vector of length 1 indicating whether or not one wants to save the ggplot object as a pdf as well, on top of the png version of the ggplot object which is produced in any case. This is an optional argument, with TRUE as default value.}

\item{lowColor}{Character vector of length 1 indicating the color of the lowest value of the scale of the 'predictions' argument. This argument will only be taken into account if a valid value is also provided for the 'highColor' argument, and when the 'predictions' argument is not considered to be a factor. This is an optional argument, with NULL as default value.}

\item{highColor}{Character vector of length 1 indicating the color of the highest value of the scale of the 'predictions' argument. This argument will only be taken into account if a valid value is also provided for the 'lowColor' argument, and when the 'predictions' argument is not considered to be a factor. This is an optional argument, with NULL as default value.}

\item{predAsFact}{Logical vector of length 1 indicating whether or the 'predictions' argument needs to be considered as a continuous or a factor variable. This is an optional argument, with FALSE as default value.}

\item{predFactColors}{Character vector of length equal to the number of different values of the 'predictions' argument. Only if the 'predAsFact' argument is toggled to TRUE, this argument is taken into account. This is an optional argument, with NULL as default value.}

\item{randomPick}{Logical vector of length 1 indicating whether one wants to let a random color generator decide upon the color of the different levels of the 'zoneGrouped' factor column of the 'inputDT' argument. If this argument is set to be TRUE, this function will generate an output object that should be used as an input for the first argument of the 'randomColorPick' function (of this package). This is an optional argument, with FALSE as default value.}
}
\value{
A ggplot object is constructed with a continuous color scheme, plotting the exposure or the predictions for the different INS regions. Note that the 'inputDT' argument needs to contain an 'exposure ' and 'postalCode' column in order to the function to work.
}
\description{
Constructing a ggplot-styled map of the Belgium, splitted up by INS region, to represent a zone by means of a continuous variable or simply the exposure
}
\examples{
library(data.table)
dirShape <- '/home/robin/shapeFiles/'
postalCode <- read.csv(paste(dirShape, "postins.csv", sep = "\\\\"))[,c("CODPOSS","INS")]
inputDT <- data.table(postalCode = postalCode$CODPOSS, exposure = runif(nrow(postalCode)))
predictions <- as.numeric(as.character(postalCode$CODPOSS))
plottingSpatialEffect(inputDT)
plottingSpatialEffect(inputDT, predictions, dirShape = dirShape)
plottingSpatialEffect(inputDT, predictions, lowColor = 'red', highColor = 'green', dirShape = dirShape)
plottingSpatialEffect(inputDT, predictions, lowColor = 'green', highColor = 'purple', dirShape = dirShape)

predictions <- sample(LETTERS[1:10], nrow(postalCode), replace = TRUE)
plottingSpatialEffect(inputDT, predictions, predAsFact = TRUE, dirShape = dirShape)
}
