% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUnivSpline.R
\name{plotUnivSpline}
\alias{plotUnivSpline}
\title{Plot the univariate spline of a GAM model}
\usage{
plotUnivSpline(
  gamFit,
  selectedVar,
  min_max_values,
  simultaneousCI = FALSE,
  nSamples = 1000,
  transparence = 0.5,
  colorLine = "black",
  colorRibbon = "steelblue",
  lineWidth = 1.2,
  showPlot = TRUE,
  plotDir = NULL,
  plotName = NULL,
  pdf2 = TRUE,
  xLimits = NULL,
  yLimits = NULL
)
}
\arguments{
\item{gamFit}{fitted GAM model, from the package [mgcv]}

\item{selectedVar}{single character variable}

\item{simultaneousCI}{logical, if the CI should be simulated}

\item{nSamples}{numeric, number of points to be used to extract spline}

\item{transparence}{alpha value for the ribbon, see [ggplot::geom_ribbon]}

\item{colorLine}{colorname or hexcode for the splice}

\item{colorRibbon}{colorname or hexcode for the ribbon}

\item{lineWidth}{linewidth for the spline}

\item{showPlot}{logical, if the plot should be shown or not}

\item{plotDir}{path location for the directory}

\item{plotName}{name of the plot, will be saved at location `plotDir`}

\item{pdf2}{logical, if the plot should be saved as a pdf}

\item{xLimits, yLimits}{numeric of length 2, indicating the limits of the plots}

\item{inputDT}{data.table object which is used to train the GAM model}
}
\value{
ggplot object of the spline
}
\description{
Plot the univariate spline of a GAM model
}
