% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splineExtract.R
\name{splineExtract}
\alias{splineExtract}
\title{Extract the spline for a variable of a GAM}
\usage{
splineExtract(
  gamFit,
  selectedVar,
  min_max_values,
  nSamples = 1000,
  simultaneousCI = FALSE
)
}
\arguments{
\item{gamFit}{gam object}

\item{selectedVar}{name of the variable from the data to create the GAM, which is a spline}

\item{min_max_values}{`numeric(2)` with the minimum and maximum values of `selectedVar`.}

\item{nSamples}{positive integer, 
the number of splits}

\item{simultaneousCI}{logical, 
if the confidence interval should be simulated}
}
\value{
data.table with columns `xVar`, `yVar`, `upperbound` and `lowerbound`
`xVar` are the values of the selected variable after transformation for the GAM 
`yVar` the spline results 
`upperbound` and `lowerbound` are the boundaries of the Confidence Interval
}
\description{
Extract the spline for a variable of a GAM
}
