% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpSuite-package.R
\docType{package}
\name{tpSuite-package}
\alias{tpSuite}
\alias{tpSuite-package}
\title{tpSuite: Technical Pricing Suite}
\description{
A collection of functions to visualize spatial data, spline effects of GAM models and how to perform binning of categorical variables for GLM models
}
\keyword{internal}
