% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform2BinnedVar.R
\name{transform2BinnedVar}
\alias{transform2BinnedVar}
\title{Categorization of continuous according to some splitting points.}
\usage{
transform2BinnedVar(inputDT, splits)
}
\arguments{
\item{inputDT}{data.table object containing the data of interest. 
This is an obligatory argument, without default value.}

\item{splits}{Named list of numeric vectors
The names of this object need to correspond to column names of the 'inputDT' argument. 
This is an obligatory argument, without default value.}
}
\value{
The variables whose names are found in the 'splits' argument as names of the object that is passed to this argument, 
are splitted by means of the splitting points that are supplied by the same 'split' argument. 
The resulting variable is added to the object that is passed to the 'inputDT' argument, 
and the respective column name corresponds to the variable name augmented with the suffix 'Grouped', 
without any space or character in between. 
Note that the grouping is always performed in the following manner: 
if the 'split' argument supplies the splitting points a, b, and c, then the continuous variable is split up in the following intervals ]-infinity, a[, [a, b[ and [b, +infinity[.
}
\description{
Categorization of continuous according to some splitting points.
}
