% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y2xSplits.R
\name{y2xSplits}
\alias{y2xSplits}
\title{Convert the y-splits into x-splits}
\usage{
y2xSplits(gamFit, selectedVar, min_max_values, ySplits)
}
\arguments{
\item{gamFit}{gam object}

\item{selectedVar}{column name for the splits}

\item{min_max_values}{`numeric(2)` with the minimum and maximum number of the selected variable}

\item{ySplits}{the splits for the y-direction}
}
\value{
splits for in the x-direction. 
The length will be greater or equal to the length of the ySplits
}
\description{
Convert the y-splits into x-splits
}
